backend: Replace gcm-server dependency with latest source
The previous gcm-server dependency was provided by theganyo, which
was an old snapshot of Google's provided gcm-server code. Since
theganyo is not keeping the library updated, replace it with a
fresh copy of Google's code. There does not appear to be a gradle
dependency provided, so the code was directly copied into this
project. The project is available on Github at:
https://github.com/google/gcm

The source code is equivalent to the October 26, 2015 commit:
a650a61165ccd5a5ecafad9683c26b6c7514a240

Hopefully an updated gradle dependency will eventually be provided.

The new features needed are topic messaging and message priority.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>