app: Workaround issue where Firebase stops receiving new data. Fixes #19
I suspect an issue with the Firebase library, so for now just force
Firebase to disconnect when the activity is paused.

The app's background data was observed to be quite high, so Firebase
is likely trying to keep the connection alive all the time. This change
should stop that behavior and save data.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>